Funcionou a impressora

+++ Back-End/src/main/java/com/ecoprint/control_center/repository/ImpresoraRepository.java
@@ -3,7 +3,7 @@ package com.ecoprint.control_center.repository;
 import com.ecoprint.control_center.model.Impressora;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImpresoraRepository extends JpaRepository<Impressora, Long>{
+public interface ImpresoraRepository extends JpaRepository<Impressora, Integer>{
